fit command history bug in ps1
When navigating up and back through command history, bash would repaint
the current line incorrectly because of something off with the escape
sequences on the coloured '$'. I'm sure there would have been a proper
way to fix it, but I'm happy enough to just move the exit code
indication to a number on the previous line.